libc: arm64: Use SIMD-accelerated memcpy
It turns out bionic is not using SIMD variant of memcpy although NEON can
be used in userspace. Import memcpy-advsimd from optimized-routines and
use it. Guard was added for extra safety but it is probably superfluous
on every modern Android device.

Initial commit in arm repo stated performance increase of 11%.

Signed-off-by: Andrzej Perczak <linux@andrzejperczak.com>
Change-Id: I757971b9996a4f5151825e0a8471880f14e06f9d